+++ examples/streaming.py
@@ -26,7 +26,7 @@ def main() -> None:
     n_actions = game.action_space().n
 
     def infer(obs: np.ndarray) -> np.ndarray:
-        return np.zeros((len(obs), 1, n_actions), dtype=np.float64)
+        return np.zeros((len(obs), 1, n_actions), dtype=np.float32)
 
     with engine.collect_stream(collect_size=args.records, infer=infer, depth=1) as stream:
         for update in range(args.updates):
